+++ sdk/appconfiguration/azure-spring-cloud-appconfiguration-config/src/main/java/com/azure/spring/cloud/config/State.java
@@ -2,8 +2,7 @@
 // Licensed under the MIT License.
 package com.azure.spring.cloud.config;
 
-import java.util.Calendar;
-import java.util.Date;
+import java.time.Instant;
 import java.util.List;
 
 import com.azure.data.appconfiguration.models.ConfigurationSetting;
@@ -12,20 +11,17 @@ class State {
 
     private final List<ConfigurationSetting> watchKeys;
 
-    private final Date nextRefreshCheck;
+    private final Instant nextRefreshCheck;
     
     private final String key;
 
     State(List<ConfigurationSetting> watchKeys, int refreshInterval, String key) {
         this.watchKeys = watchKeys;
-        Calendar calendar = Calendar.getInstance();
-        calendar.setTime(new Date());
-        calendar.add(Calendar.SECOND, refreshInterval);
-        nextRefreshCheck = calendar.getTime();
+        nextRefreshCheck = Instant.now().plusSeconds(refreshInterval);
         this.key = key;
     }
     
-    State(State oldState, Date newRefresh, String key) {
+    State(State oldState, Instant newRefresh, String key) {
         this.watchKeys = oldState.getWatchKeys();
         this.nextRefreshCheck = newRefresh;
         this.key = key;
@@ -41,7 +37,7 @@ class State {
     /**
      * @return the nextRefreshCheck
      */
-    public Date getNextRefreshCheck() {
+    public Instant getNextRefreshCheck() {
         return nextRefreshCheck;
     }