+++ sdk/appconfiguration/azure-spring-cloud-appconfiguration-config/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.4.0-beta.1 (Unreleased)
 
-* Added refresh interval parameter to `spring.cloud.azure.appconfiguraiton.config` to force refreshes on a given interval. Can be used to make sure secrets are kept up to date.
+* Added refresh interval parameter to `spring.cloud.azure.appconfiguraiton` to force refreshes on a given interval. Can be used to make sure secrets are kept up to date.
 * Added CalculatedBackoffTime, which sets the next refresh period to sooner if a refresh fails.
 
 ### Other Changes